% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's shortest path algoritm}
\usage{
dijkstra(graph, init_node, mpath = FALSE)
}
\arguments{
\item{v1}{Start node}

\item{v2}{End node}

\item{w}{Weighted path between start node and end node}
}
\value{
A vector contains the shortest paths from specific start node to other nodes
}
\description{
Dijkstra's algorithm to find the shortest path in a graph
}
