% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidean}
\alias{euclidean}
\title{Create Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A Numeric value}

\item{b}{A Numeric value}
}
\value{
The greatest common divisor
}
\description{
Euclidean algoritm to find the greatest common divisor of two integers
}
