% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidian}
\alias{euclidian}
\title{Create Euclidian algorithm}
\usage{
euclidian(a, b)
}
\arguments{
\item{a}{A Numeric value}

\item{b}{A Numeric value}
}
\value{
The greatest common divisor
}
\description{
Euclidian algoritm to find the greatest common divisor of two integers
}
