\name{eucstar-package}
\alias{eucstar-package}
\alias{eucstar}
\docType{package}
\title{
\packageTitle{eucstar}
}
\description{
\packageDescription{eucstar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{eucstar}
\packageIndices{eucstar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{eucstar}

Maintainer: \packageMaintainer{eucstar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
