% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucstar_package.R
\docType{package}
\name{eucstar}
\alias{eucstar}
\alias{eucstar-package}
\title{Euclidean and Dijkstra Package}
\description{
Contains two famous algorithms, Euclidian algoritm to find the greatest common divisor of two integers and Djikstra's shortest path algorithm.
}
\author{
Ahmet Akdeve \email{ahmak554@student.liu.se}

Zhixuan Duan \email{darinstu999@gmail.com}
}
