% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's shortest path algoritm}
\usage{
dijkstra(graph, init_node, mpath = FALSE)
}
\arguments{
\item{graph}{Data frame containing variables related to the graph}

\item{init_node}{Start node}

\item{mpath}{Multiple paths}
}
\value{
A vector contains the shortest paths from specific start node to other nodes
}
\description{
Dijkstra's algorithm to find the shortest path in a graph
}
